add search and install redux
