+++ src/App.css
@@ -39,6 +39,8 @@ section {
   color: black;
   display: flex;
   align-items: center;
+  padding-right: 0.5rem;
+  padding-left: 0.5rem;
 }
