+++ src/App.js
@@ -1,11 +1,11 @@
 import { Outlet } from 'react-router-dom';
 import './App.css';
-import NavBar from './components/NavBar';
+import NavBarPage from './components/NavBarPage';
 
 function App() {
   return (
     <div>
-      <NavBar />
+      <NavBarPage />
       <Outlet />
     </div>
   );
