+++ src/index.js
@@ -3,11 +3,11 @@ import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import App from "./App";
-import AboutStore from "./components/AboutStore";
+import AboutPage from "./components/AboutPage";
 import CatalogPage from "./components/catalog/CatalogPage";
-import Contacts from "./components/Contacts";
+import ContactsPage from "./components/ContactsPage";
 import HomePage from "./components/HomePage";
-import NotFound from "./components/NotFound";
+import NotFoundPage from "./components/NotFoundPage";
 import { store } from "./store";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
@@ -19,9 +19,9 @@ root.render(
           <Route path="/" element={<App />}>
             <Route path="/" element={<HomePage />} />
             <Route path="/catalog" element={<CatalogPage />} />
-            <Route path="/about" element={<AboutStore />} />
-            <Route path="/contacts" element={<Contacts />} />
-            <Route path="*" element={<NotFound />} />
+            <Route path="/about" element={<AboutPage />} />
+            <Route path="/contacts" element={<ContactsPage />} />
+            <Route path="*" element={<NotFoundPage />} />
           </Route>
         </Routes>
       </BrowserRouter>
