+++ src/utils.js
@@ -1,0 +1,1 @@
+export const getIsSubmit = (ev, handleSubmit) => ev.code === "Enter" ? handleSubmit(ev.target.value) : null;
